arm64: dts: Reapply "Remove unused conflicting disp_rdump_memory region"
This region is the same as the early splash framebuffer initialized by
XBL. msm_drm uses it for ramdump when available, but that's a debugging
feature we don't need. It also wastes memory after early boot because
msm_drm will keep it reserved rather than freeing the framebuffer memory
once it has taken control of the display hardware.

Remove the display ramdump region to save ~35 MiB of memory and silence
the following error on boot:

[    0.000000] OF: reserved mem: OVERLAP DETECTED!\x0adisp_rdump_region@5c000000 (0x000000005c000000--0x000000005cf00000)
overlaps with cont_splash_region@5c000000 (0x000000005c000000--0x000000005cf00000)

This reverts commit 3ed21c7f5a93ef57b5421555169bb575845bd6ce.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: mrsrimar22 <mrsrimar22@gmail.com>